[restructure]: project dirs
