+++ frontend/src/App.js
@@ -1,6 +1,6 @@
 import RemittancePage from './pages/RemittancePage';
 import {ChakraProvider, ColorModeScript, extendTheme} from '@chakra-ui/react'
-import Theme from "./theme";
+import Theme from "./components/Theme/theme";
 
 const App = () => {
     return (
